requirements-test.txt
django>=1.8
coverage
coveralls
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0

# Additional test requirements go here
bootstrap_admin
django-dynamic-preferences
tinycss